% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/la_graph.R
\name{la_graph}
\alias{la_graph}
\title{Signed digraph of the community matrix}
\usage{
la_graph(
  J,
  fname = "community",
  layout_g = "neato",
  coord = NA,
  save_file = "pdf",
  colors_nodes = NULL,
  colors_edges = NULL,
  col_option = NA
)
}
\arguments{
\item{J}{the square matrix describing the qualitative direct
causal effects between the variables of the community matrix.
See \code{Note} for more details.}

\item{fname}{the name of the \code{.dot} file generated.
Default name is \code{community}. See \code{Details} for
more information.}

\item{save_file}{defines the extension for saving the image file:
\code{"pdf"} (default) or \code{"svg"}}

\item{color}{selects the color option to visualize the signed
digraph: \code{"bw"} (default), \code{"color"}, or
\code{"greyscale"}. See \code{Details} for more information.}
}
\value{
The \code{.dot} file of the signed digraph corresponding to
the community matrix \code{J} is created and saved. Such file
can be opened with the command \code{DiagrammeR::grViz}. Also
the image of the signed digraph in either \code{.pdf} (default)
or \code{.svg} format is generated and saved using the same name
of the \code{.dot} file.
}
\description{
\code{cm_graph} visualizes the community matrix \code{J} as a
signed digraph
}
\details{
\itemize{ \item \code{fname} should be used to indicate
the name of the \code{.dot} file created. It should be noted that
the suffix \code{"_cm_graph.dot"} is automatically added to the
file name. See \code{Note} for information and \code{Examples}
for practical use; \code{color} allows the user to define the
color option for graphical visualization. The \code{color}
argument defines which options are available. The default option
(\code{"bw"}) renders a black and white signed digraph. Other
options are the use of random colors (\code{color}) and greyscale
(\code{"greyscale"}). See \code{Examples} for practical use.}
}
\note{
\itemize{ \item The matrix \code{J} should be formatted
as follows: \itemize{ \item qualitative direct causal links
must be from \code{i} (rows) variables to \code{j} (columns)
variables; \item direct causal links takes the following
coefficients: 1 for positive interactions, -1 for negative
interactions, and 0 for no interaction; \item the positive
effect of the variable \code{i} on the variable \code{j} is
represented by an arrow-headed link from \code{i} to \code{j};
\item the negative effect of the variable \code{i} on the
variable \code{j} is represented by a circle-headed link
from \code{i} to \code{j}.} \item If the argument \code{fname}
is not specified the default name \code{"community"} is assigned.
\item This function is a modification of the function
\code{\link{graph.cm}} from the R package \code{\link{LoopAnalyst}}.
This function was created on 22-11-2017.}
}
\examples{
## To save "community.pdf" image of the signed
## digraph for the community matrix M using the default color
## scheme ("bw"); the signed digraph is also saved as
## "community_cm_graph.dot"
cm_graph(M)
DiagrammeR::grViz("community_cm_graph.dot")

## To visualize the signed digraph of M with the greyscale
## color scheme and save both ".dot" and ".pdf" files with
## the name "graphM"
cm_graph(M, fname = "graphM", color = "greyscale")
DiagrammeR::grViz("community_cm_graph.dot")

## To save the colored version of the signed digraph of M
## as ".svg" file, using default name (i.e. "community")
cm_graph(M, color = "color", save_file = "svg")
DiagrammeR::grViz("community_cm_graph.dot")

}
\references{
\itemize{ \item Our manual \item
  \href{https://CRAN.R-project.org/package=LoopAnalyst}{\code{LoopAnalyst -
  CRAN}} \item \href{https://alexisdinno.com/LoopAnalyst/}{LoopAnalyst -
  Software}}
}
\seealso{
\code{\link{pred_graph}}, \code{\link{M}},
\code{\link{DiagrammeRsvg}}, \code{\link{LoopAnalyst}},
\code{\link[LoopAnalyst]{graph.cm}}, \code{\link{rsvg}}
}
\author{
Daniel Pereira \email{dsldlf@unife.it}, Alexis Dinno
\email{alexis.dinno@pdx.edu}, Antonio Bodini, Marco Scotti
\email{marcoscot@gmail.com}
}
\concept{dot, levins, loop, network, path, pdf, svg}
\keyword{array,}
\keyword{color,}
\keyword{dplot,}
\keyword{graphs,}
\keyword{hplot,}
\keyword{interface,}
\keyword{print}
