% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/la_interactions.R
\name{la_interactions}
\alias{la_interactions}
\title{Summary of simple paths}
\usage{
la_interactions(J, PL = "all", NtN = c(NA, NA), fname = NA)
}
\arguments{
\item{J}{the square matrix describing the qualitative direct
causal effects between the variables of the community matrix.
See \code{Note} for more details.}

\item{PL}{selects the path lengths for which results have to
be returned: \code{"all"} (default), \code{"sum"}, or vector
with integer, strictly positive numbers shorter than the
maximum path length in the graph. See \code{Details} for
further information.}

\item{NtN}{the vector that indicates the nodes from which
(first element) and/or to which (second element) simple
paths should be considered. See \code{Details} for additional
information.}

\item{fname}{the name of the file where the results are stored.
No file is exported when NA (default) is selected.}
}
\value{
List of matrices that summarize the number of pathways from \code{j}
  (row) nodes to \code{i} (column) nodes; each matrix corresponds to pathways
  of different lengths; plus a sum matrix of all interactions; note that it
  counts only \code{simple paths}, that is, paths that pass only once on each
  node.

  If \code{fname} is indicated, the results will be exported to a
  \code{fname.txt} file.

  An \code{invisible} list with the results is generated. It can be assigned
  to an object.
}
\description{
\code{cm_interactions} returns the summary of the simple paths
between all pairs of nodes in the graph corresponding to the
community matrix \code{J}.
}
\details{

}
\note{
\itemize{ \item The \code{J} matrix should be formatted as follows:
  \itemize{ \item should be written by indicating the qualitative direct
  casual effect from j (rows) variables on i (columns) variables; \item the
  signs of qualitative direct casual effect allowed are: \code{1} for
  positive, \code{-1} for negative and \code{0} for null.} \item Arguments
  \code{NtN} and \code{PL} can't be used together, i.e, if one or two
  variables are passed on the argument \code{NtN}, then all the information
  will be retrieved for those nodes, passing the argument \code{PL} as
  \code{"all"} even if this is changed by the user in the function call; \item See
  \code{Examples} for practical use.}
}
\examples{
## To calculate the interaction level of all nodes of M without printing the results to a .txt file
cm_interaction(M, NtN = c(), PL = "all", fname = NA)
## or (to assign the invisible results list)
results<-cm_interaction(M, NtN = c(), PL = "all", fname = NA)

## To calculate the interaction level of all nodes of M, printing the results to a "fname.txt" file
cm_interaction(M, NtN = c(), PL = "all", fname = "M")

## To calculate the interaction level of all paths begining in the node "A"
cm_interaction(M, NtN = c("A"), PL = "all", fname = NA)

## To calculate the interaction level of all paths begining in the node "A" and ending in node "E"
cm_interaction(M, NtN = c("A", "E"), PL = "all", fname = NA)

## To calculate the interaction level of all paths of length 4
cm_interaction(M, NtN= c(), PL = 4, fname = NA)

## To calculate the sum matrix of the interaction level of all nodes of M
cm_interaction(M, NtN= c(), PL = "sum", fname = NA)

}
\references{
\itemize{ \item Our manual; \item Puccia, C. J. and Levins, R.
  (1986) Qualitative Modeling of Complex Systems: An Introduction to Loop
  Analysis and Time Averaging. Cambridge: Harvard University Press; \item
  Levins, R. and Schultz, B. B. (1996) Effects of density dependence,
  feedback and environmental sensitivity on correlations among predators,
  prey and plant resources: Models and practical implications. Journal of
  Animal Ecology, 65(6),802-812.}
}
\seealso{
\code{\link{cm_paths}}, \code{\link{cm_structure}}, \code{\link{J}},
  \code{\link{igraph}}, \code{\link[igraph]{all_simple_paths}},
  \code{\link[igraph]{graph_from_adjacency_matrix}}
}
\author{
Daniel Pereira \email{dsldlf@unife.it}, Antonio Bodini, Marco Scotti
  \email{marcoscot@gmail.com}
}
\concept{interaction, levins, loop, network, path, summary}
\keyword{arith,}
\keyword{array,}
\keyword{attribute,}
\keyword{list,}
\keyword{math,}
\keyword{print}
