% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/la_paths.R
\name{la_paths}
\alias{la_paths}
\title{Community Network Paths Information}
\usage{
la_paths(J, SM, type_info = "all", path_info = NA, fname = "community")
}
\arguments{
\item{J}{a square matrix describing the qualitative direct casual effect
between the variables of a Community Network. See \code{Note} for more
details.}

\item{SM}{a square matrix with the interaction strength between the variable
of the Community Network matrix \code{J}. See \code{Note} for more
details.}

\item{type_info}{selects which path info (if any) on path information should
be in place for the printing of the paths paths information: \code{"all"}
(default), \code{"P"}, \code{"N"}, \code{"PN"}, \code{"VarC"},
\code{"StrgP"}, \code{"NodetN"}. See \code{Details} for more information on
use.}

\item{fname}{the name for the exporting file. Indication of parameter is
obligatory. See \code{Details} for more information on use.}
}
\value{
It prints a list of each path within the selected type_info option
  indictating the variables of the path, its strength, the number of nodes
  and its sign to \code{fname.txt} file.

  An \code{invisible} list with the results is generated. It can be assigned
  to an object.
}
\description{
Generates a list describing each \code{simple path} from the Community
Network matrix \code{J} with their length, sign and strength information.
}
\details{
\itemize{ \item Through the argument \code{type_info} the function
  allows the user to choose the path information to be returned
  for the printing, as follows: \itemize{ \item \code{"all"} indicates
  that there is no type_info in place and the info from all paths should be
  printed (default); \item \code{"P"} indicates that the info to be printed
  should be only from the paths with positive sign; \item \code{"N"}
  indicates that the info to be printed should be only from the paths with
  negative sign; \item \code{"PN"} indicates that the both the info from
  paths with positive and negative sign should be printed (in to separated
  files); \item \code{"VarC"} indicates that the info to be printed should be
  only from paths of higher, equal or lower number of variable as indicated
  in the  \code{VarC} vector. The vector should be constructed in the form
  \code{c("Up", x)}, as follows: \itemize{ \item \code{"Up", "Eq" or "Dn"}
  first element of the vector indicating that the print should be for paths
  with variable number higher, equal or below to the value indicated in the
  second element; \item \code{x} second element indicating the number of
  variables of the path.} \item \code{"StrgP"} indicated that the info to be
  printed should be only from paths of higher, equal or lower path strength
  as indicated in the \code{StrgP} vector. The vector should be constructed
  in the form \code{c("Up", x)}, as follows: \itemize{ \item \code{"Up" or
  "Dn"} first element of the vector indicating that the print should be for
  paths with strength higher or below to the value indicated in the second
  element; \item \code{x} second element indicating the strength of the
  path.} \item \code{"NodetN"} indicates that the info to be printed should
  be only from paths begin and/or ending in the nodes indicated in the
  \code{NodetN} vector. The vector should be constructed in the form
  \code{c(x, y)}, as follows: \itemize{ \item \code{x} first element of the
  vector indicating that the print should be for paths begining in the
  variable indicated; \item \code{x} second element of the vector indicating
  that the print should be for paths ending in the variable indicated.} \item
  see \code{Examples} for practical use.} \item \code{fname} should be used
  giving the name intended for the exporting files noting that the following
  suffixes will be added to the file name: \itemize{ \item
  \code{"_path_info.txt"} for \code{type_info
  = "all"}; \item \code{"_path_info_sign_p.txt"} for \code{type_info = "P"};  \item
  \code{"_path_info_sign_n.txt"} for
  \code{type_info = "N"};  \item \code{"_path_info_sign_p.txt"},
  \code{"_path_info_sign_n.txt"} for
  \code{type_info = "PN"};  \item \code{"_path_info_above_varc.txt"} for \code{type_info = "VarC"} where
  the first element of the vector \code{VarC} is \code{Up}; \item
  \code{"_path_info_equal_varc.txt"} for \code{type_info = "VarC"} where
  the first element of the vector \code{VarC} is \code{Eq}; \item
  \code{"_path_info_below_varc.txt"} for \code{type_info = "VarC"} where
  the first element of the vector \code{VarC} is \code{Dn}; \item
  \code{"_path_info_above_strength.txt"} for \code{type_info = "StrgP"}
  where the first element of the vector \code{StrgP} is \code{Up}; \item
  \code{"_path_info_below_strength.txt"} for \code{type_info = "StrgP"}
  where the first element of the vector \code{StrgP} is \code{Dn};\item
  \code{"_path_info_begin_nodetn.txt"} for \code{type_info = "NodetN"}
  where the first element of the vector \code{NodetN} is a variable of the
  Community Network matrix \code{J} and the second element is 0;\item
  \code{"_path_info_end_nodetn.txt"} for \code{type_info = "NodetN"} where
  the first element of the vector \code{NodetN} is 0 and the second element
  is a variable of the Community Network matrix \code{J}; \item
  \code{"_path_info_begin_end_nodetn.txt"} for \code{type_info = "NodetN"}
  where the both elements of the vector \code{NodetN} are variables of the
  Community Network matrix \code{J}; \item see \code{Note} for more
  information and \code{Examples} for practical use.}}
}
\note{
\itemize{ \item The \code{J} matrix should be formatted as follows:
  \itemize{ \item should be written by indicating the qualitative direct
  casual effect from j (rows) variables on i (columns) variables; \item The
  signs of qualitative direct casual effect allowed are: \code{1} for
  positive, \code{-1} for negative and \code{0} for null.} \item The
  \code{SM} matrix should be formatted as follows: \itemize{ \item Should be
  written by indicating the qualitative direct casual effect from j (rows)
  variables on i (columns) variables; \item The strength values between each
  variable should be within the interval \code{[0,1]}, where \code{0}
  indicates no interaction, and \code{1} indicates maximum interaction
  strength.} \item Regarding the argument \code{fname}, if not specified the
  function will be stopped and an \code{error message} will be shown
  indicating to specify \code{fname} in the function call.}
}
\examples{
## To print the information from all paths of M, printing the results to "fname.txt" and "fname.xlsx" files
cm_paths(M, SM, type_info = "all", fname = "M")
## or (to assign the invisible results list)
results <- cm_paths(M, SM, type_info = "all", fname = "M")

## To print the information from paths with positive sign
cm_paths(M, SM, type_info = "P", fname = "M")

## To print the information from paths with more than 5 variables
VarC <- c("Up", 3)
cm_paths(M, SM, type_info = "VarC", fname = "M")

## To print the information from paths with strength lower than 0.5
StrgP <- c("Dn", 0.5)
cm_paths(M, SM, type_info = "StrgP", fname = "M")

## To print the information from paths begining in node "A" and ending in node "F"
NodetN <- c("A", "E")
cm_paths(M, SM, type_info = "NodetN", fname = "M")



}
\references{
\itemize{ \item Our manual; \item and something from Loop
  analysis}
}
\seealso{
\code{\link{cm_interaction}}, \code{\link{cm_strength}},
  \code{\link{cm_structure}}, \code{\link{M}}, \code{\link{SM}},
  \code{\link{igraph}}, \code{\link[igraph]{all_simple_paths}},
  \code{\link[igraph]{graph_from_adjacency_matrix}}
}
\author{
Daniel Pereira \email{dsldlf@unife.it}, Antonio Bodini, Marco Scotti
  \email{marcoscot@gmail.com}
}
\concept{length, levins, network, paths, sign, strength, summary,}
\keyword{arith,}
\keyword{array,}
\keyword{attribute,}
\keyword{list,}
\keyword{print}
