% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/la_stability.R
\name{la_stability}
\alias{la_stability}
\title{Community Network Stability}
\usage{
la_stability(J)
}
\arguments{
\item{J}{a square matrix describing the qualitative direct casual effect
between the variables of a Community Ntwork. See \code{Note} for more
details.}
}
\value{
A print indicating the result of the test: if the Community Network matrix \code{J} is
  stable or unstable.
}
\description{
Tests the stability of the Community Network.
}
\details{
The function tests if the Community Network matrix  \code{J} is stable by
  verifying if all eigen values are negative. See \code{Note} for detail.
}
\note{
\itemize{ \item The \code{J} matrix should be formatted as follows:
  \itemize{ \item should be written by indicating the qualitative direct
  casual effect from j (rows) variables on i (columns) variables; \item the
  signs of qualitative direct casual effect allowed are: \code{1} for
  positive, \code{-1} for negative and \code{0} for null.} \item Regarding
  Community Network matrix stability: \itemize{ \item one requisite for  Community Network matrix stability is
  the sign of the the determinant. For matrices wih even number of variables,
  the \code{determinant} must be positive, while for matrices with odd number of
  variables the \code{determinant} must be negative; \item a second requisite for
  stability is for all \code{eigen values} to be negative; for this to be
  true in case of even numbered matrices the \code{determinant} must be
  positive and in case of odd numbered matrices the \code{determinant} must be negative.}}
}
\examples{
## If M is stable:
## cm_stability(M)
## [1] The community matrix is stable

## If M is not stable
## cm_stability(M)
## [1] The Community Network is unstable

}
\references{
\itemize{ \item Our manual; \item Puccia, C. J. and Levins, R. (1986) Qualitative Modeling of
  Complex Systems: An Introduction to Loop Analysis and Time Averaging.
  Cambridge: Harvard University Press.}
}
\seealso{
\code{\link{cm_structure}}, \code{\link{J}}
}
\author{
Stefania Favilla (check with Antonio if its just this), Antonio Bodini
}
\concept{network, stability}
\keyword{array,}
\keyword{math}
