% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/la_strength.R
\name{la_strength}
\alias{la_strength}
\title{Average strength of the Jacobian matrix interactions}
\usage{
la_strength(
  J,
  RM = "uniform",
  INT_MIN = NA,
  INT_MAX = NA,
  NT = 1000,
  fname = NA
)
}
\arguments{
\item{J}{the square matrix describing the qualitative direct
causal effects between the variables of the community matrix.
See \code{Note} for more details.}

\item{RM}{the parameter that allows selecting the distribution
from which the coefficients are randomly sampled and assigned
to interactions during simulation. Possible values: \code{"uniform"}
(default), \code{"normal"}, or \code{"pareto"}.
See \code{Details} for more information on use.}

\item{INT_MIN}{the matrix used to define the lower value that should
be considered by the function when randomly generating the strength
of specific interactions during simulation. The default option is
\code{NA}, which indicates that the minimum strength attained by
each non-zero interaction is \code{1e-6}.}

\item{INT_MAX}{the matrix used to define the upper value that should
be considered by the function when randomly generating the strength
of specific interactions during simulation. The default option is
\code{NA}, which indicates that the maximum strength attained by
each non-zero interaction is \code{1}.}

\item{NT}{the number number of random matrices that are generated by
the function. The default value is 1000.}

\item{fname}{the name of the file where the results are stored.
No file is exported when NA (default) is selected.}
}
\value{
Total number of matrices generated (i.e. the total number
of runs performed), the total number of stable matrices (i.e.
those effectively used for calculating the average interaction
strengths), the probability distribution used (i.e. \code{"uniform"},
\code{"normal"} or \code{"pareto"}), and the matrix with average
interaction strength for each non-zero interaction in the community
matrix \code{J}. If \code{fname} is a user-defined name of type
\cite{character} the results are saved in the file named
\code{"fname_cm_strength.txt"}. The output also consists of an
\code{invisible} list that can be can be assigned to an object.
}
\description{
\code{cm_strength} returns the matrix of the average interaction
strength based on simulation. Values are provided for all
non-zero interactions of the community matrix \code{J}.
}
\details{
\itemize{ \item Through the argument \code{RM} the user can
define the type of distribution to be used to extract the values of
intensity for the coefficient of interaction between variable
\code{i} and variable \code{j} of the community matrix \code{J}.
Options are: \itemize{\item \code{"uniform"} coefficients are
extracted from uniform probability distribution (default);
\item \code{"normal"} coefficients are extracted from (truncated)
normal probability distribution with mean = 0.5 and sd = 0.2;
\item \code{"pareto"} coefficientsare extracted from Pareto
(i.e. power-law) probability distribution with minimum value;
minimum value (location) = 1 and exponent (shape) = 8.8; \item
see \code{Note} for more information and \code{Examples} for
practical use}
\item The arguments \code{INT_MIN} and \code{INT_MAX} have
\code{NA} as default option. In such case all simulations are
performed by randomly sampling values in the interval [1e-6, 1],
according to the distribution type indicated with the argument
\code{RM}. Alternative minimum and maximum thresholds can be
changed by the user but must always lie in the interval [1e-6, 1].
\code{INT_MIN} and \code{INT_MAX} must be explicitly
provided only if the minimum or maximum values of specific
interactions are different from default. User-defined thresholds
must be in square matrices that have the sime size as the
community matrix \cite{J}. If a value is modified
for a specific interaction in \code{INT_MIN} it does not imply
that the corresponding entry must be changed in \code{INT_MAX}
(and viceversa). All zero entries in the matrices used as
input for the arguments \code{INT_MIN} and \code{INT_MAX}
indicate that default thresholds should be adopted
during simulation (i.e. minimum = 1e-6, maximum = 1). If the
user intends to define an exact interaction strength for
a specific interaction the same value should be used in the
corresponding entry of both matrices \code{INT_MIN} and
\code{INT_MIN}. See \cite{Note} for more details and
\cite{Examples} for practical use. \item The interaction
strength of the Jacobian matrix interactions are quantified
over \cite{NT} * \cite{n} runs (where \cite{n} is the size
of the Jacobian matrix \cite{J}). In each run a matrix
with randomly assigned interaction strengths is created
and it stability verified. Stable matrices only are
considered to calculate the average strength of interactions.
The argument \cite{NT} has 1000 as default value, but the user
can modify the number of times simulations are carried out
per each variable (i.e. \code{number of runs = size of J * NT}).
See \cite{Examples} for practical use. \item \code{fname} should
 be provided by the user to export the results in a text file
 with the suffix \code{"_cm_strength.txt"}. Default option is that
no text files are saved. See \code{Examples} for pratical use.}
}
\note{
\itemize{ \item The matrix \code{J} should be formatted
as follows: \itemize{ \item qualitative direct causal links
must be from \code{i} (rows) variables to \code{j} (columns)
variables; \item direct causal links takes the following
coefficients: 1 for positive interactions, -1 for negative
interactions, and 0 for no interaction; \item the positive
effect of the variable \code{i} on the variable \code{j} is
represented by an arrow-headed link from \code{i} to \code{j};
\item the negative effect of the variable \code{i} on the
variable \code{j} is represented by a circle-headed link
from \code{i} to \code{j}.} \item The \code{RM} probability
distribution from which strength values are randomly sampled
can be \code{uniform}, \code{normal} (with mean = 0.5 and
sd = 0.2; such distribution is trunctated, meaning that values
outside the renge [1e-6, 1] are discarded), or \code{pareto}
(i.e. power-law distribution) with minimum value (i.e. location)
 = 1 and exponent (i.e. shape) = 8.8. \item The \code{INT_MIN}
and \code{INT_MAX} matrices: \itemize{ \item should be written
with non-zero entries when minimum and/or maximum thresholds
do not correspond to default values (i.e. 0 and 1,
respectively); \item the value of all interaction strengths
must lie in the interval [1e-6, 1] and user-defined thresholds
only apply to matrices generated with \cite{RM = "uniform"}
(i.e. \code{INT_MIN} and \code{INT_MAX} do not work with
\cite{"normal"} and \cite{"pareto"} distributions; in such case
the function runs the analysis by arbitrarily setting
\cite{RM = "uniform"})}}
}
\examples{
## To compute the average interaction strengths of M using uniform
## distribution without saving the results in the text file
cm_strength(M)

## To assign the invisible list with all results (in presence of
## community matrix M and all default options
M_strength <- cm_strength(M)
M_strength

## To compute average interaction strengths for the community
## matrix M using the uniform distribution and saving the
## results in the file "M_cm_strength.txt"
cm_strength(M, fname = "M")

## To compute the average interaction strength of the community
## matrix M by randomly sampling the strength of each interaction
## from the truncated normal distribution (mean = 0.5, sd = 0.2)
cm_strength(M, RM = "normal")

## To compute the average interaction strength of the community
## matrix M by randomly sampling the strength of each interaction
## from the Pareto distribution
cm_strength(M, RM = "pareto")

## The average interaction strengths in case of user-defined limits
##
## (1) If the user knows that in the community matrix M the
## interaction from node A to node B can attain the minimum value
## of 0.5, then the matrix M_min has to be explicitly provided
## for the argument INT_MIN
M_min <- matrix(c(0, 0.5, 0, 0, 0,
0, 0, 0, 0, 0,
0, 0, 0, 0, 0,
0, 0, 0, 0, 0,
0, 0, 0, 0, 0), nrow = 5, byrow = TRUE)
colnames(M_min) <- colnames(M)
rownames(M_min) <- rownames(M)
cm_strength(J = M, INT_MIN = M_min)
##
## (2) In the same way, if the user knows that in the community
## matrix M the interaction strength from node C to node D can
## attain the maximum (absolute) value of 0.8, then the matrix
## M_max has to be explicitly provided for the argument INT_MAX
ro <- nrow(M)
el <- nrow(M)^2
M_max <- matrix(rep(0, el), nrow = ro)
colnames(M_max) <- colnames(M)
rownames(M_max) <- rownames(M)
M_max["C","D"] <- 0.8
cm_strength(J = M, INT_MAX = M_max)

## When average interaction strengths are computed knowing the
## exact (absolute) value of one interaction in the community
## matrix M (i.e. the effect of D on E is 0.4), then the
## following code must be used:
ro <- nrow(M)
el <- nrow(M)^2
matrix_min <- matrix(rep(0, el), nrow = ro)
colnames(matrix_min) <- colnames(M)
rownames(matrix_min) <- rownames(M)
matrix_min["D", "E"]<- 0.4
matrix_max <- matrix_min
cm_strength(M, INT_MIN = matrix_min, INT_MAX = matrix_max)

## To compute the average interaction strengths associated to
## the community matrix M and using Pareto distribution with
## user-defined number of runs (i.e. 100 for each variable, for
## a total of 500 runs):
cm_strength(M, RM = "pareto", NT = 100)

}
\references{
\itemize{ \item Our Manual
\item Martone, R.G., Bodini, A.,
Micheli, F., 2017. Identifying potential consequences of natural
perturbations and management decisions on a coastal fishery
social-ecological system using qualitative loop analysis.
Ecology and Society 22, 34.
\url{https://doi.org/10.5751/ES-08825-220134}
\item Bodini, A., Rocchi, M.
Scotti, M., 2018. Insights into the ecology of the Black Sea through
the qualitative loop analysis of the community structure.
Limnology and Oceanography 63, 968-984.
\url{https://doi.org/10.1002/lno.10713}}
}
\seealso{
\code{\link{cm_paths}}, \code{\link{cm_stability}},
  \code{\link{levins_prediction}}, \code{\link{levins_concurrent}},
  \code{\link{J}}
}
\author{
Daniel Pereira \email{dsldlf@unife.it}, Marta Rocchi, Stefania
  Favilla, Antonio Bodini, Marco Scotti \email{marcoscot@gmail.com}
}
\concept{interaction, levins, loop, network, strength}
\keyword{arith,}
\keyword{array,}
\keyword{datagen,}
\keyword{distribution,}
\keyword{math,}
\keyword{print}
