% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/la_structure.R
\name{la_structure}
\alias{la_structure}
\title{Community Network Structure}
\usage{
la_structure(J, paths = "n", fname = NA)
}
\arguments{
\item{J}{a square matrix describing the qualitative direct casual effect
between the variables of a Community Network. See \code{Note} for more
details.}

\item{fname}{the name for the exporting file. By default is \code{NA} and no
file is exported. See \code{Details} for more information on use.}
}
\value{
It returns a print of the Community Network matrix \code{J} and of
  the value for the following strutural properties : No. of Nodes of the
  Community Matrix, Total no. of Nodal Interactions, No. of Nodal
  Interactions minus self-loops, Total no. of Positive Interactions, No. of
  Positive Interactions minus self-loops, Total no. of Negative Interactions,
  No. of Negative Interactions minus self-loops, No. of Self-loop
  interactions, Number of possible paths of the Community Network matrix
  \code{J}, Number of paths crossing each variable of the Community Network
  matrix \code{J}, Trophic Level, Determinant of the Community
  Network matrix.

  If \code{fname} is indicated, the results will be exported to a
  \code{fname.txt} file.

  An \code{invisible} list with the results is generated. It can be assigned
  to an object.
}
\description{
Computes and enumerates the value of multiple structural properties from the
Community Network matrix \code{J}.
}
\details{
\code{fname} should be used giving the name intended for the
  exporting files noting that a suffix \code{"_cm_strt.txt"} will be added to
  the file name. See \code{Examples} for pratical use.
}
\note{
\itemize{ \item The \code{J} matrix should be formatted as follows:
  \itemize{ \item should be written by indicating the qualitative direct
  casual effect from j (rows) variables on i (columns) variables; \item the
  signs of qualitative direct casual effect allowed are: \code{1} for
  positive, \code{-1} for negative and \code{0} for null.} \item One
  requisite for Community Network matrix stability is the sign of the the
  \code{determinant}. For matrices wih even number of variables, the
  \code{determinant} must be positive, while for matrices with odd number of
  variables the \code{determinant} must be negative.}
}
\examples{
## To calculate the structural indices of M without printing the results to a .txt file
cm_structure(M, fname = NA)
## or (to assign the invisible results list)
results<-cm_structure(M)

## To calculate the structural indices of M, printing the results to a "fname.txt" file
cm_structure(M, fname = "M")

}
\references{
\itemize{ \item Our manual; \item Puccia, C. J. and Levins, R.
  (1986) Qualitative Modeling of Complex Systems: An Introduction to Loop
  Analysis and Time Averaging. Cambridge: Harvard University Press; \item
  Levins, R. and Schultz, B. B. (1996) Effects of density dependence,
  feedback and environmental sensitivity on correlations among predators,
  prey and plant resources: Models and practical implications. Journal of
  Animal Ecology, 65(6),802-812}.
}
\seealso{
\code{\link{cm_interaction}}, \code{\link{cm_paths}},
  \code{\link{cm_stability}}, \code{\link{J}}, \code{\link{igraph}},
  \code{\link[igraph]{all_simple_paths}},
  \code{\link[igraph]{graph_from_adjacency_matrix}}
}
\author{
Daniel Pereira \email{dsldlf@unife.it}, Antonio Bodini, Marco Scotti
  \email{marcoscot@gmail.com}
}
\concept{interaction, levins, loop, network, path, stability}
\keyword{array,}
\keyword{attribute,}
\keyword{math,}
\keyword{print}
