% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levins_concurrent.R
\name{levins_concurrent}
\alias{levins_concurrent}
\title{Levins Concurrent Prediction}
\usage{
levins_concurrent(
  J,
  DF = NA,
  RM = "uniform",
  INT_MIN = NA,
  INT_MAX = NA,
  NT = 1000,
  fname = NA
)
}
\arguments{
\item{J}{a square matrix describing the qualitative direct casual effect
between the variables of a community network. See \code{Note} for more
details.}

\item{DF}{vector containing the sign and nodes upon which the effect of
concurrent input over the Community Network matrix \code{J} is to be
investigated. See \code{Details} for more information on use.}

\item{RM}{the method to be used for random distribution for the interaction
strength values: \code{"uniform"} (default), \code{"normal"},
\code{"pareto"}. See \code{Details} for more information on use.}

\item{INT_MIN}{matrix defining the minimum value to be used by random
distribution assignment of interaction strength values of each variable
pair. The default value \code{NA} defines the minimum interaction strength
value as \code{1e-6}. See \code{Details} for more information on use..}

\item{INT_MAX}{matrix defining the maximum value to be used for the random
distribution assignment of interaction strength values of each variable
pair. The default value \code{NA} defines the maximum interaction strength
value as \code{1}. See \code{Details} for more information on use.}

\item{NT}{defines the number of random interaction strength matrices to be
randomly generated by the function. The default value is \code{1000}. See
\code{Details} for more details.}

\item{fname}{the name for the exporting file. By default is \code{NA} and no
file is exported. See \code{Details} for more information on use.}
}
\value{
It returns the total number of random matrices generated, number of
 stable matrices used for the calculations and Levins Table of Concurrent
 Predictions.

 If \code{fname} is indicated, the results will be exported to a
 \code{fname.txt} file.

 An \code{invisible} list with the results is generated. It can be assigned
 to an object.
}
\description{
Computes a table of predictions of the direction of change of the state of the
Community Network matrix \code{J} variables due to concurrent effects of
multiple inputs in variables of the same matrix.
}
\details{
\itemize{ \item Through the argument \code{DF} the function allows
 the user to indicate the direction and input nodes of a perturbation which
 will be used to compute the table of predictions of the direction of change
 in the state of the Community Network matrix \code{J} variables. The
 function allows this through the use of a vector as follows: \itemize{ \item
 odd elements indicate the sign of the input. Should be indicated as \code{1}
 for positive input and \code{-1} for negative input; \item even elements
 indicate the variable of the Community Network matrix \code{J} upon which
 the input is acting; \item e.g. \code{DF = c(1,"A", -1, "E")}; \item see
 \code{Examples} for pratical use.} \item Through the argument \code{RM} the
 user can define the type of random distribution method to be used on the
 random assignment of values of interaction strength for the variable pairs
 of the Community Matrix \code{J}, by choosing the appropriate option as
 follows: \itemize{ \item \code{"uniform"} random assignment method follows a
 Uniform probability distribution (default); \item \code{"normal"} random
 assignment method follows a Normal probability distribution; \item
 \code{"pareto"} random assignment method follows a Pareto probability
 distribution; \item see \code{Note} for more information and \code{Examples}
 for practical use.} \item Through the use of matrices for the arguments
 \code{INT_MIN} and \code{INT_MAX} the user can change the default interval
 values for the random assignment ([1e-6, 1]). This can be done by use of
 square matrices of the same length and writing format as the Community
 Network matrix \code{J}, as follows: \itemize{ \item \code{INT_MIN} the
 user defined values for each variable pair will be used by the process as
 the minimum value in the interval for random assignment ([x; 1]); \item
 \code{INT_MAX} the user defined values for each variable pair will be used
 by the process as the maximum value in the interval for random assignment
 ([1e-6; x]); \item the user defined values in \code{INT_MIN} and
 \code{INT_MAX} should be within the interval (0, 1]; \item to define an
 exact interaction strength value for any given variable pair (and not an
 interval), the user should indicate same value in both matrices; \item for
 variable pairs which the user does not pretend to change the default
 interval values, fill with value \code{0}; \item see \code{Note} for more
 information and \code{Examples} for practical use.} \item To compute a
 \code{interaction Strength Matrix} the program will compute \code{x} number
 of random matrices which will then be screened for their stability, with
 those accepeted being used to compute an average interaction strength
 matrix. Through the use of the argument \code{NT} the user can define the
 number of runs of the process. The user defined number of runs follows the
 equation: \code{Number of Runs = (Nº of variables of J * NT)}. See
 \code{Examples} for pratical use; \item \code{fname} should be used giving
 the name intended for the exporting files noting that a suffix
 \code{"_levins_conc_pred.txt"} will be added to the file name. See
 \code{Examples} for pratical use.}
}
\note{
\itemize{ \item The \code{J} matrix should be formatted as follows:
 \itemize{ \item should be written by indicating the qualitative direct
 casual effect from j (rows) variables on i (columns) variables; \item the
 signs of qualitative direct casual effect allowed are: \code{1} for
 positive, \code{-1} for negative and \code{0} for null.} \item Regarding the
 \code{INT_MIN} and \code{INT_MAX} matrices: \itemize{ \item the matrices
 should be written by indicating the interaction strength of the direct
 casual effect from j (rows) variables on i (columns) variables; \item it
 should be noted that these arguments only work with the argument \code{ RM =
 "uniform"}. If a user defined matrix is passed on either \code{INT_MIN} or
 \code{INT_MAX} and a different option other than the default is chosen for
 the argument \code{RM}, the interval considered for the value assignment
 will be the default [1e-6; 1].}}
}
\examples{
## To compute the table of predictions for two positive concurrent effects on "A" and "E", without printing the results to a .txt file
levins_concurrent(M, DF = c(1,"A", 1, "E"), RM = "uniform", INT_MIN = NA, INT_MAX = NA, NT = 100, fname = NA)
## or (to assign the invisible results list)
results <- levins_concurrent(M, DF = c(1,"A", 1, "E"), RM = "uniform", INT_MIN = NA, INT_MAX = NA, NT = 1000, fname = NA)

## To compute the table of predictions for two positive concurrent effects on "A" and "E", printing the results to a .txt file
levins_concurrent(M, DF = c(1,"A", 1, "E"), RM = "uniform", INT_MIN = NA, INT_MAX = NA, NT = 100, fname = "M")

## To compute the table of predictions for three positive concurrent effects on "A", "C" and "E", printing the results to a .txt file
levins_concurrent(M, DF = c(1,"A", 1, "C", 1, "E"), RM = "uniform", INT_MIN = NA, INT_MAX = NA, NT = 100, fname = "M")

## To compute the table of predictions for two opposite signed concurrent effects on "A" and "E"
levins_concurrent(M, DF = c(1,"A", -1, "E"), RM = "uniform", INT_MIN = NA, INT_MAX = NA, NT = 100, fname = NA)

## To compute the table of predictions for two opposite signed concurrent effects on "A" and "E", using a Normal distribution
levins_concurrent(M, DF = c(1,"A", -1, "E"), RM = "normal", INT_MIN = NA, INT_MAX = NA, NT = 100, fname = NA)

## To compute the table of predictions for two opposite signed concurrent effects on "A" and "E", using a Pareto distribution
levins_concurrent(M, DF = c(1,"A", -1, "E"), RM = "pareto", INT_MIN = NA, INT_MAX = NA, NT = 100, fname = NA)

## To compute the table of predictions for two opposite signed concurrent effects on "A" and "E", using a user defined minimum strength interval value
matrix_min<-matrix(rep(0,(length(M))), nrow = sqrt(length(M)))
matrix_min[1, 5]<- 0.3
levins_concurrent(M, DF = c(1,"A", -1, "E"), RM = "pareto", INT_MIN = matrix_min, INT_MAX = NA, NT = 100, fname = NA)

## To compute the table of predictions for two opposite signed concurrent effects on "A" and "E", using a user defined maximum strength interval value
matrix_max<-matrix(rep(0,(length(M))), nrow = sqrt(length(M)))
matrix_max[1, 5]<- 0.7
levins_concurrent(M, DF = c(1,"A", -1, "E"), RM = "pareto", INT_MIN = NA, INT_MAX = matrix_max, NT = 100, fname = NA)

## To compute the table of predictions for two opposite signed concurrent effects on "A" and "E", where [1, 5] pair has user defined fixed strength value
matrix_min<-matrix(rep(0,(length(M))), nrow = sqrt(length(M)))
matrix_min[1, 5]<- 0.4
matrix_max<-matrix(rep(0,(length(M))), nrow = sqrt(length(M)))
matrix_max[1, 5]<- 0.4
levins_concurrent(M, DF = c(1,"A", -1, "E"), RM = "pareto", INT_MIN = matrix_min, INT_MAX = matrix_max, NT = 100, fname = NA)

## To compute the table of predictions for two opposite signed concurrent effects on "A" and "E" with user defined number of runs
levins_concurrent(M, DF = c(1,"A", -1, "E"), RM = "uniform", INT_MIN = NA, INT_MAX = NA, NT = 300, fname = NA)

}
\references{
\itemize{ \item Our manual; \item and something from Loop
 analysis}
}
\seealso{
\code{\link{cm_strength}}, \code{\link{cm_stability}},
 \code{\link{levins_prediction}}, \code{\link{M}}, \code{\link{pred_graph}}
}
\author{
Daniel Pereira \email{dsldlf@unife.it}, Marta Rocchi, Antonio Bodini,
 Marco Scotti \email{marcoscot@gmail.com}
}
\concept{concurrent, interaction, levins, loop, prediction, network, sign,
 strength}
\keyword{arith,}
\keyword{array,}
\keyword{distribution,}
\keyword{math,}
\keyword{print}
